+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Bug fix in `IterableQueue` when using a multiprocessing SimpleQueue.
 - `mpservice.mpserver.Server.stream` gets a new parameter `preprocess`.
+- New classes `CyclicProcess` and `CyclicProcessWorker` in `mpservice._streamer`.
 
 
 ## [0.15.8] - 2024-0921